fix: satisfy bounded operation quality gates
